fix : 添加任务,查询返回分数

+++ src/main/java/com/ayundao/controller/WorkController.java
@@ -497,6 +497,9 @@ public class WorkController extends BaseController {
     @PostMapping("/delIndicator")
     public JsonResult delIndicator(String id) {
         //需要完善
+        if (indicatorService.find(id)==null){
+            return JsonResult.notFound("找不到指标");
+        }
         return indicatorService.delete(id, jsonResult);
     }
